# openapi.yaml
openapi: 3.1.0
info:
  title: Gozne
  version: 0.1.0-dev.4
  description:
    Alpha EVM/SIWS authentication and signed simulation controls. API and login
    require the same HTTPS origin. Authentication returns 503 without a policy.
paths:
...
  /v1/auth/nonce:
    post:
      summary: Issue a five-minute login challenge and browser context cookie
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  nonce:
                    type: string
                    pattern: ^[0-9a-f]{32}$
                  message:
                    type: string
                  expiresAt:
                    type: integer
                  signInInput:
                    type: object
                    description:
                      Sign-In With Solana fields; present only for Solana
                    additionalProperties: true
                required:
                  - nonce
                  - message
                  - expiresAt
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid proof or session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Origin or application denied, or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Missing policy or unavailable storage; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Policy changed during issuance; request a new challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                application:
                  type: string
                  maxLength: 64
                  pattern: ^[a-z][a-z0-9-]*$
                network:
                  type: string
                  enum:
                    - evm
                    - solana
                address:
                  type: string
                  maxLength: 64
                chainId:
                  type: string
                  maxLength: 32
                  description: 'Examples: 1, solana:mainnet, solana:devnet'
              required:
                - application
                - network
                - address
                - chainId
      parameters:
        - in: header
          name: Origin
          required: true
          schema:
            type: string
            format: uri
          description: Exact application HTTPS origin
  /v1/auth/verify:
    post:
      summary:
        Verify the exact message and issue a session lasting at most one hour
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid proof or session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Origin or application denied, or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Missing policy or unavailable storage; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                nonce:
                  type: string
                  pattern: ^[0-9a-f]{32}$
                message:
                  type: string
                  maxLength: 4096
                signature:
                  type: string
                  maxLength: 256
                  description:
                    'EVM: 65 hexadecimal bytes prefixed with 0x. Solana: 64
                    bytes in base64.'
              required:
                - nonce
                - message
                - signature
      security:
        - LoginContext: []
      parameters:
        - in: header
          name: Origin
          required: true
          schema:
            type: string
            format: uri
          description: Exact application HTTPS origin
  /v1/auth/me:
    get:
      summary: Read current session and CSRF token
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid proof or session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Origin or application denied, or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Missing policy or unavailable storage; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - SessionCookie: []
  /v1/auth/validate:
    get:
      summary: Internal proxy validation; do not expose to browsers
      responses:
        '200':
          description: Success
          headers:
            X-Gozne-Identity:
              schema:
                type: string
              description: Internal identity
            X-Gozne-Role:
              schema:
                type: string
              description: Comma-separated roles
            X-Gozne-Application:
              schema:
                type: string
              description: Authorized application
            X-Gozne-Session:
              schema:
                type: string
              description: Public audit ID, not a session token
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid proof or session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Origin or application denied, or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Missing policy or unavailable storage; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - SessionCookie: []
      parameters:
        - in: query
          name: application
          required: true
          schema:
            type: string
            maxLength: 64
            pattern: ^[a-z][a-z0-9-]*$
  /v1/auth/logout:
    post:
      summary: Revoke the session and clear cookies
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  ok:
                    type: boolean
                    const: true
                required:
                  - ok
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid proof or session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Origin or application denied, or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Missing policy or unavailable storage; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - SessionCookie: []
      parameters:
        - in: header
          name: Origin
          required: true
          schema:
            type: string
            format: uri
          description: Exact application HTTPS origin
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
            pattern: ^[0-9a-f]{64}$
...
components:
  securitySchemes:
    LoginContext:
      type: apiKey
      in: cookie
      name: __Host-gozne-login
      description: Secure; HttpOnly; SameSite=Strict; Path=/; no Domain
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-gozne-session
      description: Secure; HttpOnly; SameSite=Strict; Path=/; no Domain
  schemas:
    Session:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        identity:
          type: string
        application:
          type: string
        roles:
          type: array
          items:
            type: string
        expiresAt:
          type: integer
          description: Unix milliseconds
        csrfToken:
          type: string
          pattern: ^[0-9a-f]{64}$
        network:
          type: string
          enum:
            - evm
            - solana
        address:
          type: string
      required:
        - id
        - identity
        - application
        - roles
        - expiresAt
        - csrfToken
        - network
        - address
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              format: uuid
          required:
            - code
            - message
            - request_id
      required:
        - error
...